feat: add more business logic to friend request service

+++ src/main/java/com/techeer/f5/jmtmonster/domain/friend/service/FriendRequestService.java
@@ -2,7 +2,10 @@ package com.techeer.f5.jmtmonster.domain.friend.service;
 
 import com.techeer.f5.jmtmonster.domain.friend.dao.FriendRequestRepository;
 import com.techeer.f5.jmtmonster.domain.friend.domain.FriendRequest;
+import com.techeer.f5.jmtmonster.global.error.exception.DuplicateResourceException;
+import com.techeer.f5.jmtmonster.global.error.exception.FieldErrorWrapper;
 import com.techeer.f5.jmtmonster.global.error.exception.ResourceNotFoundException;
+import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -18,33 +21,56 @@ public class FriendRequestService {
 
     @Transactional
     public FriendRequest create(FriendRequest entity) {
-        // TODO: Check duplicate
-        // TODO: Check validity of users
+        checkExistsByFromUserIdAndToUserId(entity);
+
         return friendRequestRepository.save(entity);
     }
 
     @Transactional
     public FriendRequest update(UUID id, FriendRequest updatedEntity) {
-        // TODO: Check and get entity using ID
-        // TODO: Check duplicity
-        // TODO: Check validity of users
-        // TODO: Check actual queries
-        return friendRequestRepository.save(updatedEntity);
+        checkExistsByFromUserIdAndToUserId(updatedEntity);
+
+        FriendRequest existingEntity = findOneById(id);
+        existingEntity.update(
+                updatedEntity.getFromUser(),
+                updatedEntity.getToUser(),
+                updatedEntity.getStatus()
+        );
+        return friendRequestRepository.save(existingEntity);
     }
 
     @Transactional
     public void deleteById(UUID id) {
+        if (!friendRequestRepository.existsById(id)) {
+            throw new ResourceNotFoundException(FriendRequest.class.getSimpleName(), "id", id);
+        }
         friendRequestRepository.deleteById(id);
     }
 
     @Transactional(readOnly = true)
     public FriendRequest findOneById(UUID id) {
         return friendRequestRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("friend-requests", "id", id));
+                .orElseThrow(() -> new ResourceNotFoundException(
+                        FriendRequest.class.getSimpleName(), "id", id));
     }
 
     @Transactional(readOnly = true)
     public Page<FriendRequest> findAll(Pageable pageable) {
         return friendRequestRepository.findAll(pageable);
+    }
+
+    @Transactional(readOnly = true)
+    public void checkExistsByFromUserIdAndToUserId(FriendRequest entity) {
+        UUID fromUserId = entity.getFromUser().getId();
+        UUID toUserId = entity.getToUser().getId();
+        String resourceName = FriendRequest.class.getSimpleName();
+
+        if (friendRequestRepository.existsByFromUser_idAndToUser_id(fromUserId, toUserId)) {
+            throw new DuplicateResourceException(resourceName,
+                    List.of(new FieldErrorWrapper(resourceName, "fromUserId", fromUserId.toString(),
+                                    "already exists with toUserId"),
+                            new FieldErrorWrapper(resourceName, "toUserId", toUserId.toString(),
+                                    "already exists with fromUserId")));
+        }
     }
 }
